Refactoring (step 3).
* Filters, Players, etc. now work on double[]s (instead of byte[]s).
* This gives us much more precision when apply filters/effects.

+++ src/com/a31morgan/sound/player/IFilter.java
@@ -2,8 +2,8 @@ package com.a31morgan.sound.player;
 
 public interface IFilter {
 
-	public void applyNoteFilter(byte[] input);
+	public void applyNoteFilter(double[] input);
 
-	public void applyStreamFilter(byte[] input);
+	public void applyStreamFilter(double[] input);
 
 }
